# Tuning gpairs implementation params